% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{are_null}
\alias{are_null}
\title{Vectorised is.null}
\usage{
are_null(x)
}
\arguments{
\item{x}{list of things that may contain NULL values}
}
\value{
logical
}
\description{
Vectorised is.null
}
\examples{
is.null(list(NULL, NULL, 1))
are_null(list(NULL, NULL, 1))
are_null(list(NULL, NULL, NULL))
are_null(list(1, 2, 3))
is.null(list(1, 2, 3))
}
