% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greta_model_class.R
\name{as.greta_model}
\alias{as.greta_model}
\title{register generic method to coerce objects to a greta model}
\usage{
as.greta_model(x, ...)
}
\arguments{
\item{x}{greta_model object}

\item{...}{extra arguments - not used.}
}
\description{
register generic method to coerce objects to a greta model
}
