% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unknowns_class.R
\name{as.unknowns}
\alias{as.unknowns}
\title{Create objects of class 'unknowns' to nicely print ? valued arrays}
\usage{
as.unknowns(x)
}
\arguments{
\item{x}{object to convert to "unknowns" class}
}
\description{
Create objects of class 'unknowns' to nicely print ? valued arrays
}
