% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chol.greta_array}
\alias{chol.greta_array}
\title{Compute the Cholesky Factor of a Matrix}
\usage{
\method{chol}{greta_array}(x, ..., force_cholesky = FALSE)
}
\arguments{
\item{x}{an object for which a method exists.  The default method
    applies to numeric (or logical) symmetric, positive-definite matrices.}

\item{...}{further arguments pass to or from methods.}

\item{force_cholesky}{Whether to force cholesky computation. Currently
used as a workaround to ensure cholesky is calculated properly, and may
result in code that uses \code{chol()} to be slow. Default is TRUE. Can change
to FALSE, but may encounter issues in
\url{https://github.com/greta-dev/greta/issues/585}.}
}
\description{
Compute the Cholesky Factor of a Matrix
}
