% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_greta_env.R
\name{create_conda_greta_env}
\alias{create_conda_greta_env}
\title{Create a special conda greta env thing

Helper function to install appropriate python packages and python version
for \code{greta}.}
\usage{
create_conda_greta_env()
}
\description{
Create a special conda greta env thing

Helper function to install appropriate python packages and python version
for \code{greta}.
}
