% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_types.R
\name{data}
\alias{data}
\title{Declare greta Arrays as Data}
\usage{
data(x)
}
\arguments{
\item{x}{an object that can be coerced to an array}
}
\description{
define an object in an R session as a greta array for use as
  data in a greta model.
}
\details{
This overwrites the \code{\link[utils]{data}} function in the
  \code{utils} package, which is used to access datasets. You can still
  access that function by doing \code{utils::data()}
}
