% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unknowns_class.R
\name{dim<-.unknowns}
\alias{dim<-.unknowns}
\title{set dims like on a matrix/array}
\usage{
\method{dim}{unknowns}(x) <- value
}
\arguments{
\item{x}{matrix/array to set values to}

\item{value}{values that are  being set set}
}
\description{
set dims like on a matrix/array
}
