% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_class.R
\name{dim.node}
\alias{dim.node}
\title{generic to grab dimensions of nodes}
\usage{
\method{dim}{node}(x)
}
\arguments{
\item{x}{greta node class}
}
\value{
An integer vector of the node's dimensions.
}
\description{
generic to grab dimensions of nodes
}
