% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{grete-distributions}
\alias{flat}
\alias{free}
\alias{grete-distributions}
\alias{lognormal}
\alias{normal}
\title{grete probability distributions}
\usage{
normal(mu, sigma, dim = 1)

lognormal(mu, sigma, dim = 1)

flat(range, dim = 1)

free(dim = 1)
}
\arguments{
\item{mu, sigma}{probability distribution parameters}

\item{dim}{the dimensions of the variable, by default a scalar}

\item{range}{a finite, length 2 numeric vector giving the range of values to
which \code{flat} distributions are constrained. The first element must
be lower than the second.}
}
\description{
These probability distributions can be used to define random
  variables in a grete model. They return a 'node' object that can be
  combined with other nodes to construct a model.
}
\details{
Most of these distributions have non-uniform probability densities,
  however the distributions \code{flat} and \code{free} do not. These can
  therefore be used as parameters in likelihood (rather than Bayesian)
  inference. By default, all distributions are
}
\examples{
# a fixed distribution, e.g. for a prior
mu = normal(0, 1)

# an unconstrained, positive parameter sigma
log_sigma = free()
sigma = exp(log_sigma)

# a hierarchical distribution
theta = normal(mu, lognormal(0, 1))

# a vector of 3 variables drawn from the same hierarchical distribution
thetas = normal(mu, sigma, dim = 3)

# a matrix of 12 variables drawn from the same hierarchical distribution
thetas = normal(mu, sigma, dim = c(3, 4))

# a constrained parameter with no density (e.g. for a constrained likelihood model)
theta = flat(c(1, 5))
}

