% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{evaluate greta arrays given fixed values}
\usage{
evaluate(target, values)
}
\arguments{
\item{target}{a greta array for which to evaluate the value}

\item{values}{a named list giving temporary values of the greta arrays with
which \code{target} is connected.}
}
\value{
A numeric R array with the same dimensions as \code{target}, giving
  the values it would take conditioned on the fixed values given by
  \code{values}.
}
\description{
Evaluate greta arrays as numeric R arrays, given temporary
  values for the greta arrays on which they depend. This can be used to check
  the behaviour of your model or make predictions to new data after model
  fitting.
}
\details{
The greta arrays named in \code{values} need not be variables, they
  can also be other operations or even data.

  At present, \code{values} must contain values for \emph{all} of the
  variable greta arrays with which \code{target} is connected, even values are
  given for intermediate operations, or the target doesn't depend on the
  variable. That may be relaxed in a future release.
}
\examples{
\dontrun{

# define a variable greta array, and another that is calculated from it
# then evaluate what value y would take for different values of x
x = normal(0, 1, dim = 3)
a = lognormal(0, 1)
y <- sum(x ^ 2) + a
evaluate(y, list(x = c(0.1, 0.2, 0.3), a = 2))

# define a simple model
alpha = normal(0, 1)
beta = normal(0, 1)
sigma = lognormal(1, 0.1)
mu <- alpha + iris$Petal.Length * beta
distribution(iris$Petal.Width) = normal(mu, sigma)
m <- model(alpha, beta, sigma)

# work out what intermediate values would be, given some parameter values
evaluate(mu[1:5], list(alpha = 1, beta = 2))
evaluate(mu[1:5], list(alpha = -1, beta = 0.2))
}
}
