% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_replace_combine.R
\name{extract-replace-combine}
\alias{extract-replace-combine}
\alias{extract}
\alias{replace}
\alias{[.greta_array}
\alias{[<-.greta_array}
\alias{cbind.greta_array}
\alias{rbind.greta_array}
\alias{c.greta_array}
\alias{rep.greta_array}
\alias{dim.greta_array}
\alias{length.greta_array}
\alias{dim<-.greta_array}
\alias{head.greta_array}
\alias{tail.greta_array}
\alias{diag.greta_array}
\title{extract, replace and combine greta arrays}
\usage{
\method{[}{greta_array}(x, ...)

\method{[}{greta_array}(x, ...) <- value

\method{cbind}{greta_array}(...)

\method{rbind}{greta_array}(...)

\method{c}{greta_array}(...)

\method{rep}{greta_array}(x, ...)

\method{dim}{greta_array}(x)

\method{length}{greta_array}(x)

\method{dim}{greta_array}(x) <- value

\method{head}{greta_array}(x, n = 6L, ...)

\method{tail}{greta_array}(x, n = 6L, ...)

\method{diag}{greta_array}(x = 1, nrow, ncol)
}
\arguments{
\item{x}{a greta array}

\item{...}{either further indices specifying elements to extract or replace
(\code{i}, \code{j} for \code{[}), or multiple greta arrays to combine (\code{cbind()},
\code{rbind()} & \code{c()}), or additional arguments (\code{rep()}, \code{head()}, \code{tail()}).
Generic arguments such as \code{drop} and \code{recursive} are accepted but ignored
for greta arrays.}

\item{value}{for \verb{[<-} a greta array to replace elements, for
\verb{dim<-} either NULL or a numeric vector of dimensions}

\item{n}{a single integer, as in \code{utils::head()} and
\code{utils::tail()}}

\item{nrow, ncol}{optional dimensions for the resulting greta array when x is
not a matrix.}
}
\value{
A \code{greta_array}, with elements extracted, replaced, or combined
as appropriate for the method called.
}
\description{
Generic methods to extract and replace elements of greta arrays,
or to combine greta arrays.
}
\details{
\code{diag()} can be used to extract or replace the diagonal part of
a square and two-dimensional greta array, but it cannot be used to create a
matrix-like greta array from a scalar or vector-like greta array. A static
diagonal matrix can always be created with e.g. \code{diag(3)}, and then
converted into a greta array.

Also note that since R 4.0.0, \code{head} and \code{tail} methods for arrays changed
to print a vector rather than maintain the array structure. The \code{greta}
package supports both methods, and will do so based on which version of R
you are using.
}
\section{Syntax}{

These methods can be used with greta arrays as follows:

\if{html}{\out{<div class="sourceCode">}}\preformatted{# extract
x[i]
x[i, j, ..., drop = FALSE]
head(x, n = 6L, ...)
tail(x, n = 6L, ...)
diag(x, nrow, ncol)

# replace
x[i] <- value
x[i, j, ...] <- value
diag(x) <- value

# combine
cbind(...)
rbind(...)
abind(...)
c(..., recursive = FALSE)
rep(x, times, ..., recursive = FALSE)

# get and set dimensions
length(x)
dim(x)
dim(x) <- value
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{

x <- as_data(matrix(1:12, 3, 4))

# extract and replace
x[1:3, ]
x[, 2:4] <- 1:9
e <- diag(x)
diag(x) <- e + 1

# combine
cbind(x[, 2], x[, 1])
rbind(x[1, ], x[3, ])
abind(x[1, ], x[3, ], along = 1)
c(x[, 1], x)
rep(x[, 2], times = 3)
}
}
