% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gpu_cpu}
\alias{gpu_cpu}
\alias{gpu_only}
\alias{cpu_only}
\title{Set GPU or CPU usage}
\usage{
gpu_only()

cpu_only()
}
\description{
These functions set the use of CPU or GPU inside of greta. They
simply return either "GPU" or "CPU", but in the future may handle more
complexity. These functions arepassed to \code{compute_options} inside of a few
functions: \code{\link[=mcmc]{mcmc()}}, \code{\link[=opt]{opt()}}, and \code{\link[=calculate]{calculate()}}.
}
