% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\name{greta-inference}
\alias{greta-inference}
\alias{mcmc}
\title{Statistical Inference on Greta Models}
\usage{
mcmc(model, method = c("hmc"), n_samples = 1000, thin = 1, warmup = 100,
  verbose = TRUE, control = list(), initial_values = NULL)
}
\arguments{
\item{model}{greta_model object}

\item{method}{the method used to sample values. Currently only \code{hmc} is
implemented}

\item{n_samples}{the number of samples to draw (after any warm-up, but before
thinning)}

\item{thin}{the thinning rate; every \code{thin} samples is retained, the
rest are discarded}

\item{warmup}{the number of samples to spend warming up the sampler. During
this phase the sampler moves toward the highest density area and may tune
sampler hyperparameters.}

\item{verbose}{whether to print progress information to the console}

\item{control}{an optional named list of hyperparameters and options to
control behaviour of the sampler. See Details.}

\item{initial_values}{an optional named vector of initial values for the free
parameters in the model}
}
\value{
\code{mcmc} - an \code{mcmc.list} object that can be analysed using
  functions from the coda package. This will on contain mcmc samples of the
  parameters of interest, as defined in \code{model}.
}
\description{
Carry out statistical inference on parameters of interest by MCMC
}
\details{
If \code{verbose = TRUE}, the progress bar shows the number of iterations
  so far and the expected time to complete the phase of model fitting (warmup
  or sampling). Occasionally, a proposed set of parameters can have such a
  low density as to cause numerical instability (I.e. the log density or its
  gradient is \code{NA},\code{Inf} or \code{-Inf}). When this happens, the
  progress bar will also display the proportion of samples so far that were
  'bad' (numerically unstable) and therefore rejected. If you're getting a
  lot of these, you may want to manually define starting values to move the
  sampler into a more reasonable part of the aprameter space.

  Currently, the only implemented MCMC procedure is static Hamiltonian
  Monte Carlo (\code{method = 'hmc'}). During the warmup iterations, the
  leapfrog stepsize hyperparameter \code{epsilon} is tuned to maximise the
  sampler efficiency. The \code{control} argument can be used to specify the
  initial value for epsilon, along with two other hyprparameters: \code{Lmin}
  and \code{Lmax}; positive integers (with \code{Lmax > Lmin}) giving the
  upper and lower limits to the number of leapfrog steps per iteration (from
  which the number is selected uniformly at random).

  If you're getting a lot of rejected proposals, you might want to reduce
  \code{epsilon}, or if the sampler is moving too slowly you might want to
  increase \code{epsilon}.

  The default control options for HMC are:
  \code{control = list(Lmin = 10, Lmax = 20, epsilon = 0.005)}
}
\examples{
\dontrun{

# define a simple model
mu = free()
sigma = lognormal(1, 0.1)
x = rnorm(10)
distribution(x) = normal(mu, sigma)
m <- model(mu, sigma)

# carry out mcmc on the model
draws <- mcmc(m,
              n_samples = 100,
              warmup = 10)
}
}
