% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax.R
\name{greta-likelihood}
\alias{greta-likelihood}
\alias{likelihood<-}
\alias{likelihood}
\title{Define a Likelihood over Data}
\usage{
likelihood(data) <- value
}
\arguments{
\item{data}{either a data greta array (defined using \code{data()}), or
some data that can be coereced to a data greta array}

\item{value}{a stochastic greta array, created using a distribution}
}
\description{
The likelihood function is used to link observed data with
  random variables. This can be used to define the likelhood term for a
  model.
}
\examples{
# observed data
y = rnorm(100, 0, 3)

# mean and variance parameters (with no priors)
mu = free()
sigma = exp(free())

# define the likelihood
likelihood(y) = normal(mu, sigma)

}
