% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greta_model_class.R
\name{greta-model}
\alias{greta-model}
\alias{print.greta_model}
\title{Greta Model Objects}
\usage{
\method{print}{greta_model}(x, ...)
}
\arguments{
\item{x}{a \code{greta_model} object}

\item{\dots}{further arguments passed to or from other methods (currently ignored).}
}
\description{
Methods to inspect \code{greta_model} objects.
}
