% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{greta-operations}
\alias{greta-operations}
\title{operations on greta nodes}
\description{
This is a list of currently implemented operations to combine
 greta nodes into probabilistic models. More will follow.
}
\section{Usage}{
 \preformatted{
 x + y
 x - y
 x * y
 x / y
 log(x)
 exp(x)
 x \%*\% y
 }
}

