% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{greta-operators}
\alias{greta-operators}
\title{arithmetic operators for greta nodes}
\description{
This is a list of currently implemented arithmetic operators to
  combine greta nodes into probabilistic models.
}
\details{
greta's operators are used just like R's the standard arithmetic
  operators, but they return other greta nodes, rather than values.
}
\section{Usage}{
 \preformatted{
 # arithmetic operators
 -x
 x + y
 x - y
 x * y
 x / y
 x ^ y
 x \%\% y
 x \%/\% y
 x \%*\% y

 # logical operators
 !x
 x & y
 x | y
 }
}
\examples{
 # arithmetic
 x = observed(-1:12)
 a = x + 1
 b = 2 * x + 3
 c = x \%\% 2
 d = x \%/\% 5

 # logical
 e = (x > 1) | (x < 1)
 f = e & (x < 2)
 g = !f
}

