% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{greta-structures}
\alias{greta-structures}
\alias{zeros}
\alias{ones}
\title{greta Data Structures}
\usage{
zeros(...)

ones(...)
}
\arguments{
\item{...}{dimensions of the greta arrays to create}
}
\value{
a greta array object
}
\description{
These structures can be used to set up more complex models. For
  example, scalar parameters can be embedded in a greta array by first creating a
  greta array with \code{zeros()} or \code{ones()}, and then embedding the
  parameter value using greta's replacement syntax
}
\examples{
# a 3 row, 4 column greta array of 0s
z <- zeros(3, 4)

# a 3x3x3 greta array of 1s
z <- ones(3, 3, 3)
}
