% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{greta-transforms}
\alias{greta-inverse-links}
\alias{greta-transforms}
\alias{icauchit}
\alias{icloglog}
\alias{ilogit}
\alias{iprobit}
\alias{log1pe}
\title{transformation functions for greta nodes}
\usage{
iprobit(x)

ilogit(x)

icloglog(x)

icauchit(x)

log1pe(x)
}
\arguments{
\item{x}{a real-valued (i.e. values ranging from -Inf to Inf) greta node to
transform to a constrained value}
}
\description{
parameter transformations for greta nodes, which may also be
  used as inverse link functions
}
\details{
greta does not allow you to state the transformation/link on the
  left hand side of an assignment, as is common in the BUGS and STAN
  modelling languages. That's because the same syntax has a very different
  meaning in R, and can only be applied to objects that are already in
  existence. The inverse forms of the common link functions (prefixed with an
  'i') are therefore more likely to useful in modelling, and these are what
  are provided here.

  The log1pe inverse link function is equivalent to \code{log(1 + exp(x))},
  yielding a positive transformed parameter. Unlike the log transformation,
  this transformation is approximately linear for x > 1. i.e. when \eqn{x >
  1}, \deqn{y \approx x}
}
\examples{
 x = normal(1, 3, dim = 10)

 # transfromation to the unit interval
 p1 <- iprobit(x)
 p2 <- ilogit(x)
 p3 <- icloglog(x)
 p4 <- icauchit(x)

 # and to positive reals
 y <- log1pe(x)
}

