% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greta_package.R
\docType{package}
\name{greta}
\alias{greta}
\alias{greta-package}
\title{greta: Probabilistic Modelling with TensorFlow}
\description{
greta lets you write probabilistic models interactively in
  native R code, then sample from them efficiently using Hamiltonian Monte
  Carlo.

  The computational heavy lifting is done by TensorFlow, Google's automatic
  differentiation library. greta is particularly fast where the model
  contains lots of linear algebra, and greta models can be easily set up to
  run across CPUs or GPUs just by installing the relevant version of
  TensorFlow.

  See the example below for the general set up of a greta model, and
  \link{greta-distributions}, \link{greta-operators}, \link{greta-functions},
  \link{greta-transforms} and \link{greta-structures} for details of the
  currently implemented syntax and how to combine them into models
}
\examples{
\dontrun{
# define a simple model
mu = free()
sigma = lognormal(1, 0.1)
x = rnorm(10)
distribution(x) = normal(mu, sigma)

m <- define_model(mu, sigma)

# and sample from it
draws <- mcmc(m,
              n_samples = 100,
              warmup = 10)
}
}
