% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-deps-tf-tfp.R
\docType{data}
\name{greta_deps_tf_tfp}
\alias{greta_deps_tf_tfp}
\title{Suggested valid Python dependencies for greta}
\format{
\subsection{\code{greta_deps_tf_tfp}}{

A data frame with 63 rows and 5 columns:
\describe{
\item{os}{Operating System}
\item{tfp_version, tf_version}{numeric versions in format major.minor.patch for TFP and TF}
\item{python_version_min, python_version_max}{numeric versions range in format major.minor.patch for Python}
}
}
}
\usage{
greta_deps_tf_tfp
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This dataset is deprecated and will be removed in a future release. greta
stopped validating against it in 0.6.0, and it is a fixed snapshot that stops
at TensorFlow 2.15.1 -- well behind the versions greta now installs -- so
consulting it gives the wrong answer about what works.

For which combinations currently install and run, see the checks greta runs
weekly across Linux, macOS and Windows:
\url{https://github.com/greta-dev/greta/actions/workflows/install-check.yaml}.
}
\details{
This is a dataset that contains suggested valid versions of Tensorflow (TF),
Tensorflow Probability (TFP), and Python for linux, mac, and windows
machines. It was constructed from
\url{https://www.tensorflow.org/install/source} and
\url{https://www.tensorflow.org/install/source_windows}, and by inspecting
\url{https://github.com/tensorflow/probability/releases}.

We recommend using the default versions provided in \code{greta_deps_spec()}.
}
\keyword{datasets}
