% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greta_install_miniconda.R
\name{greta_install_miniconda}
\alias{greta_install_miniconda}
\title{Installs miniconda}
\usage{
greta_install_miniconda(timeout = 5)
}
\arguments{
\item{timeout}{time (minutes) until installation stops. Default is 5 minutes.}
}
\value{
nothing - installs miniconda.
}
\description{
This installs miniconda using \code{\link[reticulate:install_miniconda]{reticulate::install_miniconda()}} inside
\code{\link[callr:r_process_options]{callr::r_process_options()}}. Used internally by \code{\link[=install_greta_deps]{install_greta_deps()}}.
We mostly recommend users use \code{\link[=install_greta_deps]{install_greta_deps()}} to manage their
python dependency installation.
}
