% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greta-sitrep.R
\name{greta_list_py_modules}
\alias{greta_list_py_modules}
\title{List Python modules installed in greta env}
\usage{
greta_list_py_modules()
}
\value{
matrix/data frame of Python modules that are installed in the greta environment.
}
\description{
List Python modules installed in greta env
}
