% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_greta_deps.R
\name{greta_python_deps}
\alias{greta_python_deps}
\title{Specify python dependencies for greta}
\usage{
greta_python_deps(
  tf_version = "2.15.0",
  tfp_version = "0.23.0",
  python_version = "3.10"
)
}
\arguments{
\item{tf_version}{Character. Tensorflow (TF) version in format major.minor.patch. Default is "2.15.0".}

\item{tfp_version}{Character.Tensorflow probability (TFP) version major.minor.patch. Default is "0.23.0".}

\item{python_version}{Character. Python version in format major.minor.patch. Default is "3.10".}
}
\value{
list of dependencies
}
\description{
A helper function for specifying versions of Tensorflow (TF), Tensorflow Probability (TFP), and Python.
}
\examples{
greta_python_deps()
}
