% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-logfiles.R
\name{greta_set_install_logfile}
\alias{greta_set_install_logfile}
\title{Set logfile path when installing greta}
\usage{
greta_set_install_logfile(path)
}
\arguments{
\item{path}{valid path to logfile - should end with \code{.html} so you
can benefit from the html rendering.}
}
\value{
nothing - sets an environment variable for use with
\code{\link[=install_greta_deps]{install_greta_deps()}}.
}
\description{
To help debug greta installation, you can save all installation output
to a single logfile.
}
