% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greta-sitrep.R
\name{greta_sitrep}
\alias{greta_sitrep}
\title{Greta Situation Report}
\usage{
greta_sitrep(verbosity = c("minimal", "detailed", "quiet"))
}
\arguments{
\item{verbosity}{character. How verbose the output of the situation report.
Possible options: "minimal" (default), "detailed", and "quiet". "Minimal"
provides just information in python version, tensorflow version,
tensorflow probability, and whether greta conda environment is available.
"Quiet" presents no information, but prepares greta to be used. "Detailed"
gives information on the version and path for R, greta, python,
tensorflow, tensorflow probability, the greta conda environment, and a
statement on greta usability.}
}
\value{
Message on greta situation report. See "verbosity" parameter details
above for more information.
}
\description{
This checks if Python, Tensorflow, Tensorflow Probability, and the greta
conda environment are available, and also loads and initialises python
}
\examples{
\dontrun{
greta_sitrep()
}
}
