% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greta_stash.R
\name{greta_notes_install_miniconda}
\alias{greta_notes_install_miniconda}
\alias{installation-stash}
\alias{greta_notes_conda_create}
\alias{greta_notes_conda_install}
\title{Retrieve python installation details

These functions retrieve installation information output by python
when running \code{install_miniconda()}, \code{conda_create()} and \code{conda_install()}.}
\usage{
greta_notes_install_miniconda()

greta_notes_conda_create()

greta_notes_conda_install()
}
\description{
Retrieve python installation details

These functions retrieve installation information output by python
when running \code{install_miniconda()}, \code{conda_create()} and \code{conda_install()}.
}
\examples{
greta_notes_install_miniconda()
greta_notes_conda_create()
greta_notes_conda_install()
}
