% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\docType{data}
\name{internals}
\alias{internals}
\alias{.internals}
\title{internal methods}
\format{An object of class \code{list} of length 4.}
\usage{
.internals
}
\description{
\code{.internals} contains a list of functions and R6 class
  objects that can be used to develop extensions to greta. Most greta users
  will not need to access these methods, and it is not recommended to use
  \code{.internals} in model code.
}
\keyword{datasets}
