% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{is.greta_array}
\alias{is.greta_array}
\title{Is object a greta array?}
\usage{
is.greta_array(x, ...)
}
\arguments{
\item{x}{object to test if is greta array}

\item{...}{extra args (currently not used)}
}
\description{
Is object a greta array?
}
