% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greta_mcmc_list.R
\name{is.greta_mcmc_list}
\alias{is.greta_mcmc_list}
\title{Is object a \code{greta_mcmc_list}?}
\usage{
is.greta_mcmc_list(x, ...)
}
\arguments{
\item{x}{An object that may be a greta_mcmc_list}

\item{...}{extra args (not currently used)}
}
\value{
logical TRUE/FALSE
}
\description{
Is object a \code{greta_mcmc_list}?
}
