% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax.R
\name{likelihood}
\alias{\%~\%}
\alias{likelihood}
\title{create an observed, stochastic node}
\usage{
data \%~\% distribution
}
\arguments{
\item{data}{a data node, defined using \code{observed()}}

\item{distribution}{a stochastic node, created using a distribution}
}
\description{
The likelihood operator is used to link observed data with
  random variables. This can be used to define the likelhood term for a
  model.
}
\examples{
# observed data
y = observed(rnorm(10))

# random variable
theta = normal(0, 1)

# link them (i.e. we observed theta to have the values in y)
y \%~\% theta
}

