% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginalisers.R
\name{marginalisers}
\alias{marginalisers}
\alias{discrete_marginalisation}
\alias{laplace_approximation}
\title{marginalisation methods}
\usage{
discrete_marginalisation(values)

laplace_approximation(stepsize = 0.05, tolerance = 1e-06,
  max_iterations = 50)
}
\arguments{
\item{values}{an R vector giving values at which to evaluate the function for a
discrete marginalisation}

\item{stepsize}{the (positive) size of steps in the Newton-Raphson
optimisation algorithm used to find the approximate conditional posterior
mode of the random variable}

\item{max_iterations}{the (positive) integer-like maximum number of iterations of
the Newton-Raphson optimisation algorithm}

\item{stepsize}{the (positive) numerical convergence tolerance (in units of
the log conditional posterior density) of the Newton-Raphson optimisation
algorithm}
}
\value{
a \code{marginaliser} object that can be passed to
  \code{marginalise}.
}
\description{
Functions to set up marginalisers (which explicitly integrate
  out random variables) and modify their behaviour, for use in
  \code{\link{marginalise}()}.

  \code{discrete_marginalisation} marginalises discrete random variables via
  a Riemann sum over a range of plausible values.

  \code{laplace_approximation} marginalises random variables (with
  multivariate normal priors) approximately, using by Newton-Raphson
  optimisation and assuming the posterior follows a multivariate normal
  distribution.
}
\details{
\code{discrete_marginalisation} can only be used with discrete
  probability distributions, e.g. those defined with \code{poisson()} and
  \code{binomial()}. For discrete distributions with finite support (such as
  \code{bernoulli()}) the marginalisation will be exact, so long as
  \code{values} includes all possible values of the variable. For discrete
  distributions with non-finite support (such as \code{poisson()}, which has
  no upper bound), the marginalisation can only ever be approximate. However
  if \code{values} cover a range of values with sufficiently high support in
  the distribution, that approximation error will be minimal.

\code{laplace_approximation} can only be used to marginalise
  variables following a multivariate normal distribution. In addition, the
  function to be marginalised must \emph{factorise}; ie. it must return a
  vector-valued density with as many elements as the vector variable being
  marginalised, and each of element of the density must depend only on the
  corresponding element of the variable vector. This is the responsibility of
  the user, and is not checked.
}
