% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-logfiles.R
\name{open_greta_install_log}
\alias{open_greta_install_log}
\title{Read a greta logfile}
\usage{
open_greta_install_log(path = NULL)
}
\arguments{
\item{path}{file to read. Optional. If not specified, it will search for
the environment variable "GRETA_INSTALLATION_LOG". To set
"GRETA_INSTALLATION_LOG" you can use
\code{Sys.setenv('GRETA_INSTALLATION_LOG'='path/to/logfile.html')}. Or use
\code{\link[=greta_set_install_logfile]{greta_set_install_logfile()}} to set the path, e.g.,
\code{greta_set_install_logfile('path/to/logfile.html')}.}
}
\value{
opens a URL in your default browser
}
\description{
This is a convenience function to facilitate reading logfiles. It opens
a browser using \code{\link[utils:browseURL]{utils::browseURL()}}.
}
