% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greta_mcmc_list.R
\name{print.greta_mcmc_list}
\alias{print.greta_mcmc_list}
\title{Print method for greta MCMC list}
\usage{
\method{print}{greta_mcmc_list}(x, ..., n = 5)
}
\arguments{
\item{x}{greta mcmc list}

\item{...}{extra args (currently not used)}

\item{n}{number of lines to print}
}
\value{
printed MCMC output
}
\description{
Print method for greta MCMC list
}
