% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{py_module_available}
\alias{py_version}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{reticulate}{\code{\link[reticulate:py_module_available]{py_module_available()}}, \code{\link[reticulate:py_version]{py_version()}}}
}}

