% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimiser_class.R
\name{run_optimiser}
\alias{run_optimiser}
\title{Dispatch optimisation method to right class}
\usage{
run_optimiser(self)
}
\arguments{
\item{self}{optimiser of class: \code{tf_optimiser}, \code{tfp_optimiser}, or
\code{tf_compat_optimiser}.}
}
\value{
Invisibly returns \code{NULL}; called for its side effect of running
the optimiser and updating its internal state.
}
\description{
Should also allow for building other methods in the future
}
