% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greta_stash.R
\name{greta_notes_install_miniconda}
\alias{greta_notes_install_miniconda}
\alias{stash-notes}
\alias{greta_notes_conda_create}
\alias{greta_notes_conda_install}
\alias{greta_notes_tf_num_error}
\title{Retrieve python installation or error details

These functions retrieve installation or error information output by python
when running \code{install_miniconda()}, \code{conda_create()}, \code{conda_install()}, or
when encountering a TensorFlow numerical problem.}
\usage{
greta_notes_install_miniconda()

greta_notes_conda_create()

greta_notes_conda_install()

greta_notes_tf_num_error()
}
\description{
Retrieve python installation or error details

These functions retrieve installation or error information output by python
when running \code{install_miniconda()}, \code{conda_create()}, \code{conda_install()}, or
when encountering a TensorFlow numerical problem.
}
\examples{
\dontun{
greta_notes_install_miniconda()
greta_notes_conda_create()
greta_notes_conda_install()
greta_notes_tf_num_error()
greta_notes_tf_error()
}
}
