% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turn-off-autoconfig.R
\name{turn_off_autoconfig}
\alias{turn_off_autoconfig}
\title{Helper function to turn off Reticulate Autoconfigure}
\usage{
turn_off_autoconfig()
}
\value{

}
\description{
Sometimes it is required for installation of `greta` to change the .Renviron
  to turn off Reticulate's autoconfigure. This function assists the user
  in turning off autoconfigure.
}
\examples{
\dontrun{
turn_off_autoconfig()
}
}
