% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-logfiles.R
\name{write_greta_install_log}
\alias{write_greta_install_log}
\title{Write greta dependency installation log file}
\usage{
write_greta_install_log(path = greta_install_logfile())
}
\arguments{
\item{path}{a path with an HTML (.html) extension. Defaults to the
\code{GRETA_INSTALLATION_LOG} environment variable if set, otherwise
"greta-installation-logfile.html" in \code{tools::R_user_dir("greta")}.}
}
\value{
nothing - writes to file
}
\description{
This can only be run after installation has happened with
\code{\link[=install_greta_deps]{install_greta_deps()}}, and before restarting R.
}
