% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greta_model_class.R
\name{as.greta_model}
\alias{as.greta_model}
\title{Convert object to a "greta_model" object}
\usage{
as.greta_model(x, ...)
}
\arguments{
\item{x}{object to convert to greta model}

\item{...}{extra arguments}
}
\description{
Convert object to a "greta_model" object
}
