% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data}
\alias{data}
\title{Declare R Objects as Data}
\usage{
data(x)
}
\arguments{
\item{x}{an R object that can be coerced to a greta_array (see details).}
}
\description{
define an object in an R session as a data greta array for use
  as data in a greta model.
}
\details{
\code{data} can currently convert R objects to greta_arrays if they are
  numeric or logical vectors, matrices or arrays; or if they are dataframes
  with only numeric (including integer) or logical elements. Logical elements
  are always converted to numerics. R objects cannot be converted if they
  contain missing (\code{NA}) or infinite (\code{-Inf} or \code{Inf}) values.

  \code{greta::data()} masks the \code{\link[utils]{data}} function in
  the \code{utils} package, which is used to access datasets. You can still
  access that function by doing \code{utils::data()}.
}
\examples{

# numeric/integer/logical vectors, matrices and arrays can all be coerced to
# data greta arrays

vec <- rnorm(10)
mat <- matrix(seq_len(3 * 4), nrow = 3)
arr <- array(sample(c(TRUE, FALSE), 2 * 2 * 2, replace = TRUE), dim = c(2, 2, 2))
(a <- data(vec))
(b <- data(mat))
(c <- data(arr))

# dataframes can also be coerced, provided all the columns are numeric,
# integer or logical
df <- data.frame(x1 = rnorm(10),
                 x2 = sample(1L:10L),
                 x3 = sample(c(TRUE, FALSE), 10, replace = TRUE))
(d <- data(df))

}
