% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinstallers.R
\name{destroy_greta_deps}
\alias{destroy_greta_deps}
\title{Remove greta dependencies and remove miniconda}
\usage{
destroy_greta_deps()
}
\value{
nothing
}
\description{
Sometimes when installing greta you might encounter an error and the best
thing to do is start from a clean slate. This function does two things:
\enumerate{
\item Removes the "greta-tf2-env" with \code{\link[=remove_greta_env]{remove_greta_env()}}
\item Removes the miniconda installation with \code{\link[=remove_miniconda]{remove_miniconda()}}
}
}
