% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax.R
\name{distribution}
\alias{distribution}
\alias{distribution<-}
\alias{likelihood}
\alias{distribution}
\alias{likelihood<-}
\title{Define a Distribution Over a greta Array}
\usage{
distribution(greta_array) <- value

distribution(greta_array)

likelihood(greta_array) <- value
}
\arguments{
\item{greta_array}{a greta array. For the assignment method it must be a
greta array that doesn't already have a probability distribution.}

\item{value}{a greta array with a distribution (see
\code{\link{greta-distributions}})}
}
\description{
\code{distribution} links observed data, variables, and other
  greta arrays to probability distributions. For example a model likelhood
  can be set by using \code{distribution} on some observed data.
  \code{likelihood} is an alias for \code{distribution}. It is deprecated and
  will be removed in version 0.2.
}
\details{
The extract method returns the greta array if it has a distribution,
  or \code{NULL} if it doesn't. It has now real function, but is included for
  completeness

  \code{distribution} can also be used to create truncated distributions, by
  first defining a greta array with constraints (the truncation) and then
  defining the distribution on that greta array. See below for an example.
}
\examples{

# define a model likelihood

# observed data and mean parameter to be estimated
# (explicitly coerce data to a greta array so we can refer to it later)
y = as_data(rnorm(5, 0, 3))
mu = variable()
# define the distribution over y (the model likelihood)
distribution(y) = normal(mu, 1)

# get the distribution over y
distribution(y)

# define a truncated-positive standard normal random variable
tn = variable(lower = 0)
distribution(tn) = normal(0, 1)

}
