% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_replace_combine.R
\name{extract-replace-combine}
\alias{cbind,}
\alias{extract,}
\alias{extract-replace-combine}
\alias{rbind}
\alias{replace,}
\title{Extract, Replace and Combine Nodes}
\arguments{
\item{i, j}{indices specifying elements to extract or replace}

\item{value}{a node to replace elements}

\item{...}{either further indices specifying elements to extract or replace,
or multiple nodes to combine (\code{cbind} & \code{rbind})}

\item{recursive}{generic argument ignored for greta nodes}
}
\description{
Generic methods to extract and replace elements of nodes, or to
  combine nodes.
}
\section{Usage}{
 \preformatted{
# extract
x[i]
x[i, j, ...]

# replace
x[i] <- value
x[i, j, ...] <- value

# combine
cbind(...)
rbind(...)
c(..., recursive = FALSE)
}
}
\examples{
 x = observed(matrix(1:12, 3, 4))

 # extract/replace
 x[1:3, ]
 x[, 2:4] <- 1:9

 # combine
 cbind(x[, 2], x[, 1])
 rbind(x[1, ], x[3, ])
 c(x[, 1], x[, 2])

}

