% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_module.R
\name{gp-module}
\alias{gp-module}
\alias{gp}
\title{methods for Gaussian process modelling}
\arguments{
\item{variance, variances}{(scalar/vector) the variance of a Gaussian process
prior in all dimensions (\code{variance}) or in each dimensions
(\code{variances}).}

\item{lengthscale, lengthscales}{(scalar/vector) the correlation decay
distance along all dimensions (\code{lengthscale}) or each dimension
((\code{lengthscales})) of the Gaussian process}

\item{period}{(scalar) the period of the Gaussian process}

\item{dim}{(scalar integer, not a greta array) the dimension of the Gaussian
process (number of columns on which it acts)}
}
\description{
A module providing composable kernel functions for use in
  Gaussian process models. Currently the only provided methods are the kernel
  constructor functions under gp$kernels. Each of these returns a
  \emph{function} which can be executed on greta arrays to compute the
  covariance matrix between points in the space of the Gaussian process. See
  the example for a demonstration.
}
\details{
The kernels are imported from the GPflow python package, using the
  gpflowr R package. Both of those need to be installed before you can use
  these methods. See the \href{gpflow.readthedocs.io}{GPflow website} for
  details of the kernels implemented.
}
\section{Usage}{

\preformatted{
  gp$kernels$bias(variance, dim = 1)
  gp$kernels$white(variance, dim = 1)
  gp$kernels$linear(variances)
  gp$kernels$rbf(lengthscales, variance)
  gp$kernels$exponential(lengthscales, variance)
  gp$kernels$matern12(lengthscales, variance)
  gp$kernels$matern32(lengthscales, variance)
  gp$kernels$matern52(lengthscales, variance)
  gp$kernels$periodic(period, lengthscale, variance)
}
}

\examples{
# create a radial basis function kernel on two dimensions
k1 <- gp$kernels$rbf(lengthscales = c(0.1, 0.2), variance = 0.6)

# evaluate it on a greta array to get the variance-covariance matrix
x <- greta_array(rnorm(8), dim = c(4, 2))
k1(x)

# non-symmetric covariance between two sets of points
x2 <- greta_array(rnorm(10), dim = c(5, 2))
k1(x, x2)


}
