% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax.R
\name{greta-distribution}
\alias{greta-distribution}
\alias{distribution<-}
\alias{distribution}
\alias{likelihood}
\alias{likelihood<-}
\alias{distribution}
\title{Define a Distribution Over a greta Array}
\usage{
distribution(greta_array) <- value

likelihood(greta_array) <- value

distribution(greta_array)
}
\arguments{
\item{greta_array}{any greta array that doesn't already have a probability
distribution.}

\item{value}{a greta array with a probability distribution (see
\code{\link{greta-distributions}})}
}
\description{
The distribution function is used to link observed data, free
  parameters and other greta arrays with probability distributions. For
  example, \code{distribution} can be used to define the likelhood term for a
  model by using \code{distribution} on some observed data. \code{likelihood}
  is an alias for \code{distribution}. It is deprecated and will be removed
  in version 0.2.
}
\examples{
# observed data
y_ = rnorm(5, 0, 3)

# explicitly coerce to a data greta array (so we can refer to it later)
y <- as_data(y_)
y

# mean and variance parameters (with no priors)
mu = free()
sigma = exp(free())

# define the distribution over y (the model likelihood)
distribution(y) = normal(mu, sigma)

# get the distribution over y
distribution(y)

}
