% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{greta-distributions}
\alias{bernoulli}
\alias{beta}
\alias{binomial}
\alias{exponential}
\alias{flat}
\alias{free}
\alias{gamma}
\alias{greta-distributions}
\alias{lognormal}
\alias{normal}
\alias{poisson}
\alias{student}
\title{greta probability distributions}
\usage{
free(dim = 1)

normal(mu, sigma, dim = 1)

lognormal(mu, sigma, dim = 1)

bernoulli(p, dim = 1)

binomial(trials, p, dim = 1)

poisson(lambda, dim = 1)

gamma(shape, scale, dim = 1)

exponential(rate, dim = 1)

student(mean, df, dim = 1)

beta(alpha, beta, dim = 1)

flat(range, dim = 1)
}
\arguments{
\item{dim}{the dimensions of the variable, by default a scalar}

\item{mu}{unconstrained parameter}

\item{sigma}{positive parameter (\code{sigma > 0})}

\item{p}{probability parameter (\code{0 < p < 1})}

\item{trials}{positive integer parameter (\code{trials} in \code{{1, 2, 3, ...}})}

\item{lambda}{positive parameter (\code{lambda > 0})}

\item{shape}{positive parameter (\code{shape > 0})}

\item{scale}{positive parameter (\code{scale > 0})}

\item{rate}{positive parameter (\code{rate > 0})}

\item{mean}{unconstrained parameter}

\item{df}{positive parameter (\code{df > 0})}

\item{alpha}{positive parameter (\code{alpha > 0})}

\item{beta}{positive parameter (\code{beta > 0})}

\item{range}{a finite, length 2 numeric vector giving the range of values to
which \code{flat} distributions are constrained. The first element must
be lower than the second.}
}
\description{
These probability distributions can be used to define random
  variables in a greta model. They return a 'node' object that can be
  combined with other nodes to construct a model.
}
\details{
Most of these distributions have non-uniform probability densities,
  however the distributions \code{flat} and \code{free} do not. These can
  therefore be used as parameters in likelihood (rather than Bayesian)
  inference.

  The discrete probability distributions (\code{bernoulli}, \code{poisson})
  can be used as likelihoods, but not as unknown variables.
}
\examples{
# a fixed distribution, e.g. for a prior
mu = normal(0, 1)

# an unconstrained, positive parameter sigma
log_sigma = free()
sigma = exp(log_sigma)

# a hierarchical distribution
theta = normal(mu, lognormal(0, 1))

# a vector of 3 variables drawn from the same hierarchical distribution
thetas = normal(mu, sigma, dim = 3)

# a matrix of 12 variables drawn from the same hierarchical distribution
thetas = normal(mu, sigma, dim = c(3, 4))

# a constrained parameter with no density (e.g. for a constrained likelihood model)
theta = flat(c(1, 5))
}

