% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{greta-functions}
\alias{greta-functions}
\title{functions for greta nodes}
\description{
This is a list of currently implemented functions to transform
 greta nodes. More will follow.
}
\section{Usage}{
 \preformatted{
 log(x)
 exp(x)
 t(x)
 }
}

