% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{greta-functions}
\alias{greta-functions}
\title{R functions that work for greta nodes}
\description{
This is a list of functions in base R that are currently
  implemented to transform greta nodes.
}
\details{
TensorFlow only enables rounding to integers, so \code{round()} will
  error is set to anything other than \code{0}.

  Any additional arguments to \code{chol()} will be ignored, see the
  TensorFlow documentation for details of the cholesky routine.

  \code{diag()} can only be used to extract the diagonal part of a square
  matrix, it cannot be used to create a matrix from a node, nor to assign the
  diagonal elements of a square matrix. A static diagonal matrix can always
  be created with e.g. \code{diag(3)}.
}
\section{Usage}{
 \preformatted{

 # logarithms and exponentials
 log(x)
 exp(x)
 log1p(x)
 expm1(x)

 # miscellaneous mathematics
 abs(x)
 sqrt(x)
 sign(x)

 # rounding of numbers
 ceiling(x)
 floor(x)
 round(x, digits = 0)

 # trigonometry
 cos(x)
 sin(x)
 tan(x)
 acos(x)
 asin(x)
 atan(x)

 # special mathematical functions
 lgamma(x)
 digamma(x)

 # matrix operations
 t(x)
 chol(x, ...)
 diag(x, nrow, ncol)
 }
}
\examples{
x = observed(matrix(1:9, nrow = 3, ncol = 3))
a = log(exp(x))
b = log1p(expm1(x))
c = sign(x - 5)
d = abs(x - 5)

e = diag(x)

z = t(a)


}

