% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\name{greta-inference}
\alias{greta-inference}
\alias{define_model}
\alias{mcmc}
\title{Statistical Inference on Greta Models}
\usage{
define_model(...)

mcmc(model, method = c("hmc"), n_samples = 1000, thin = 1, warmup = 100,
  verbose = TRUE, control = list(), initial_values = NULL)
}
\arguments{
\item{\dots}{\code{greta_array} objects to be tracked by the model (i.e.
those for which samples will be retained during mcmc). If not provided, all
of the non-data \code{greta_array} objects defined in the calling
environment will be tracked.}

\item{model}{greta_model object}

\item{method}{the method used to sample values. Currently only \code{hmc} is
implemented}

\item{n_samples}{the number of samples to draw (after any warm-up, but before
thinning)}

\item{thin}{the thinning rate; every \code{thin} samples is retained, the
rest are discarded}

\item{warmup}{the number of samples to spend warming up the sampler. During
this phase the sampler moves toward the highest density area and may tune
sampler hyperparameters.}

\item{verbose}{whether to print progress information to the console}

\item{control}{an optional named list of hyperparameters and options to
control behaviour of the sampler}

\item{initial_values}{an optional named vector of initial values for the free
parameters in the model}
}
\value{
\code{define_model} - a \code{greta_model} object. See
  \code{\link{greta-model}} for details.

\code{mcmc} - an \code{mcmc.list} object that can be analysed using
  functions from the coda package. This will on contain mcmc samples of the
  parameters of interest, as defined in \code{model}.
}
\description{
Define a \code{greta_model} object and carry out statistical
  inference on parameters of interest by MCMC
}
\examples{

\dontrun{

# define a simple model
mu = free()
sigma = lognormal(1, 0.1)
x = rnorm(10)
distribution(x) = normal(mu, sigma)

m <- define_model(mu, sigma)

}
\dontrun{
# carry out mcmc on the model
draws <- mcmc(m,
              n_samples = 100,
              warmup = 10)
}
}
