% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_greta_deps.R
\name{greta_deps_receipt}
\alias{greta_deps_receipt}
\title{Capture greta python dependencies.}
\usage{
greta_deps_receipt()
}
\value{
\code{greta_deps_spec()} object
}
\description{
To assist with capturing and sharing python dependencies, we provide a way
to capture the dependencies currently used.
}
\examples{
\dontrun{
my_deps <- greta_deps_receipt()
}
}
