% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-deps-tf-tfp.R
\docType{data}
\name{greta_deps_tf_tfp}
\alias{greta_deps_tf_tfp}
\title{Suggested valid Python dependencies for greta}
\format{
\subsection{\code{greta_deps_tf_tfp}}{

A data frame with 63 rows and 5 columns:
\describe{
\item{os}{Operating System}
\item{tfp_version, tf_version}{numeric versions in format major.minor.patch for TFP and TF}
\item{python_version_min, python_version_max}{numeric versions range in format major.minor.patch for Python}
}
}
}
\usage{
greta_deps_tf_tfp
}
\description{
This is a dataset that contains suggested valid versions of Tensorflow (TF),
Tensorflow Probability (TFP), and Python for linux, mac, and windows
machines. It was constructed from
\url{https://www.tensorflow.org/install/source} and
\url{https://www.tensorflow.org/install/source_windows}, and by inspecting
\url{https://github.com/tensorflow/probability/releases}.
}
\details{
We recommend using the default versions provided in \code{greta_deps_spec()}.
}
\keyword{datasets}
