% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{greta_sitrep}
\alias{greta_sitrep}
\title{Greta Situation Report}
\usage{
greta_sitrep()
}
\value{
Message if greta is ready to use
}
\description{
This checks if Python, Tensorflow, Tensorflow Probability, and the greta
conda environment are available, and also loads and initialises python
}
\examples{
\dontrun{
greta_sitrep()
}
}
