% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_greta_deps.R
\name{install_greta_deps}
\alias{install_greta_deps}
\title{Install Python dependencies for greta}
\usage{
install_greta_deps(
  method = c("auto", "virtualenv", "conda"),
  conda = "auto",
  ...
)
}
\arguments{
\item{method}{Installation method ("virtualenv" or "conda")}

\item{conda}{The path to a conda executable. Use "auto" to allow reticulate
to automatically find an appropriate conda binary. See Finding Conda for
more details.}

\item{...}{Other arguments passed to \code{\link[tensorflow:install_tensorflow]{tensorflow::install_tensorflow()}}.}
}
\description{
This is a helper function to install Python dependencies needed. This
includes Tensorflow version 1.14.0, Tensorflow Probability 0.7.0, and
numpy version 1.16.4. These Python modules will be installed into an
"r-tensorflow" virtual or conda environment. Note that "virtualenv" is
not available on Windows.
}
\note{
This function was heavily inspired by \code{keras::install_keras()}.
}
\examples{
\dontrun{
install_greta_deps()
}
}
