% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_greta_deps.R, R/reinstallers.R
\name{install_greta_deps}
\alias{install_greta_deps}
\alias{reinstall_greta_deps}
\title{Install Python dependencies for greta}
\usage{
install_greta_deps(
  python_deps = greta_python_deps(),
  timeout = 5,
  restart = c("ask", "force", "no"),
  ...
)

reinstall_greta_deps(
  python_deps = greta_python_deps(),
  timeout = 5,
  restart = c("ask", "force", "no")
)
}
\arguments{
\item{python_deps}{object created with \code{greta_python_deps()} where you
specify python, TF, and TFP versions.}

\item{timeout}{maximum time in minutes until the installation for each
installation component times out and exits. Default is 5 minutes per
installation component.}

\item{restart}{character. Restart R after installation? Default is "ask".
Other options are, "force", and "no". Using "force" will will force a
restart after installation. Using  "no" will not restart. Note that this
only restarts R during interactive sessions, and only in RStudio.}

\item{...}{Optional arguments, reserved for future expansion.}
}
\description{
This is a helper function to install Python dependencies needed. This
includes the latest version of Tensorflow version (2.13.0 or higher),
Tensorflow Probability 0.21.0 (or higher), and the latest version of
numpy (1.21.0 or higher). These Python modules will be installed into a
virtual or conda environment, named "greta-env-tf2". Note that "virtualenv"
is not available on Windows.
}
\details{
By default, if using RStudio, it will now ask you if you want to restart
the R session. If the session is not interactive, or is not in RStudio,
it will not restart. You can also override this with \code{restart = TRUE}.
}
\note{
This will automatically install Miniconda (a minimal version of the
Anaconda scientific software management system), create a 'conda'
environment for greta named 'greta-env-tf2' with required python and python
package versions, and forcibly switch over to using that conda environment.

If you don't want to use conda or the "greta-env-tf2" conda environment, you
can install these specific versions of tensorflow (version 2.6.0), and
tensorflow-probability (version 0.14.1), and ensure that the python
environment that is initialised in this R session has these versions
installed. This is now always straightforward, so we recommend installing
the python packages using \code{install_greta_deps()} for most users.
}
\examples{
\dontrun{
install_greta_deps()
}
\dontrun{
# to help troubleshoot your greta installation, this can help resolve some
# issues with installing greta dependencies
reinstall_greta_deps()
}
}
