% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_greta_deps.R
\name{install_greta_deps}
\alias{install_greta_deps}
\title{Install Python dependencies for greta}
\usage{
install_greta_deps(
  method = c("auto", "virtualenv", "conda"),
  conda = "auto",
  timeout = 5,
  manual = FALSE,
  ...
)
}
\arguments{
\item{method}{Installation method ("virtualenv" or "conda")}

\item{conda}{The path to a \code{conda} executable. Use \code{"auto"} to allow
\code{reticulate} to automatically find an appropriate \code{conda} binary. See
\strong{Finding Conda} for more details.}

\item{timeout}{maximum time in minutes until the installation for each
installation component times out and exits. Default is 5 minutes per
installation component.}

\item{manual}{logical. Skip the fancy installation and just run:
\code{reticulate::py_install( packages = c( 'numpy', 'tensorflow', 'tensorflow-probability' ), pip = TRUE )}}

\item{...}{Optional arguments, reserved for future expansion.}
}
\description{
This is a helper function to install Python dependencies needed. This
includes the latest version of Tensorflow version (2.8.0 or higher),
Tensorflow Probability 0.16.0 (or higher), and the latest verison of
numpy (1.21.0 or higher). These Python modules will be installed into a
virtual or conda environment, named "greta-env-tf2". Note that "virtualenv"
is not available on Windows.
}
\note{
This will automatically install Miniconda (a minimal version of the
Anaconda scientific software management system), create a 'conda'
environment for greta named 'greta-env-tf2' with required python and python
package versions, and forcibly switch over to using that conda environment.

If you don't want to use conda or the "greta-env-tf2" conda environment, you
can install these specific versions of tensorflow (version 2.6.0), and
tensorflow-probability (version 0.14.1), and ensure that the python
environment that is initialised in this R session has these versions
installed. This is now always straightforward, so we recommend installing
the python packages using \code{install_greta_deps()} for most users.
}
\examples{
\dontrun{
install_greta_deps()
}
}
