% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginalisers.R
\name{marginalisers}
\alias{marginalisers}
\alias{discrete_marginalisation}
\title{marginalisation methods}
\usage{
discrete_marginalisation(values)
}
\arguments{
\item{values}{an R vector giving values at which to evaluate the function for a
discrete marginalisation}
}
\value{
a \code{marginaliser} object that can be passed to
  \code{marginalise}.
}
\description{
Functions to set up marginalisers (which explicitly integrate
  out random variables) and modify their behaviour, for use in
  \code{\link{marginalise}()}.
}
\details{
\code{discrete_marginalisation} can only be used with discrete
  probability distributions, e.g. those defined with \code{poisson()} and
  \code{binomial()}. For discrete distributions with finite support (such as
  \code{bernoulli()}) the marginalisation will be exact, so long as
  \code{values} includes all possible values of the variable. For discrete
  distributions with non-finite support (such as \code{poisson()}, which has
  no upper bound), the marginalisation can only ever be approximate. However
  if \code{values} cover a range of values with sufficiently high support in
  the distribution, that approximation error will be minimal.
}
