% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_types.R
\name{observed}
\alias{observed}
\title{define observed data}
\usage{
observed(data)
}
\arguments{
\item{data}{an object that can be coerced to an array}
}
\description{
define an object in an R session as data in a greta model
}

