% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overloaded.R, R/operators.R, R/functions.R
\name{overloaded}
\alias{overloaded}
\alias{\%*\%}
\alias{diag}
\alias{colMeans}
\alias{rowMeans}
\alias{colSums}
\alias{rowSums}
\alias{sweep}
\alias{backsolve}
\alias{forwardsolve}
\alias{tapply}
\title{Functions overloaded by greta}
\usage{
x \%*\% y

diag(x = 1, nrow, ncol)

colMeans(x, na.rm = FALSE, dims = 1L)

rowMeans(x, na.rm = FALSE, dims = 1L)

colSums(x, na.rm = FALSE, dims = 1L)

rowSums(x, na.rm = FALSE, dims = 1L)

sweep(x, MARGIN, STATS, FUN = "-", check.margin = TRUE, ...)

backsolve(r, x, k = ncol(r), upper.tri = TRUE, transpose = FALSE)

forwardsolve(l, x, k = ncol(l), upper.tri = FALSE, transpose = FALSE)

tapply(X, INDEX, FUN, ...)
}
\arguments{
\item{x, y, nrow, ncol, MARGIN, STATS, FUN, check.margin, \dots, r, k, upper.tri, transpose, l, na.rm, dims, X, INDEX, FUN}{arguments as in original documentation}
}
\description{
greta provides a wide range of methods to apply common R
  functions and operations to \code{greta_array} objects. A few of these
  functions and operators are not associated with a class system, so they are
  overloaded here. This should not affect normal use of these functions, but
  they need to be documented to satisfy CRAN's check.
}
