% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overloaded.R, R/operators.R, R/functions.R
\name{overloaded}
\alias{overloaded}
\alias{\%*\%}
\alias{diag}
\alias{sweep}
\title{Functions overloaded by greta}
\usage{
x \%*\% y

diag(x = 1, nrow, ncol)

sweep(x, MARGIN, STATS, FUN = "-", check.margin = TRUE, ...)
}
\arguments{
\item{x, y, nrow, ncol, MARGIN, STATS, FUN, check.margin, \dots}{arguments as in original documentation}
}
\description{
greta provides a wide range of methods to apply common R
  functions and operations to \code{greta_array} objects. A few of these
  functions and operators are not associated with a class system, so they are
  overloaded here. This should not affect normal use of these functions, but
  they need to be documented to satisfy CRAN's check.
}
