% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_greta_deps.R
\name{print.greta_deps_spec}
\alias{print.greta_deps_spec}
\title{Print method for greta python deps}
\usage{
\method{print}{greta_deps_spec}(x, ...)
}
\arguments{
\item{x}{greta python deps}

\item{...}{extra args, not used}
}
\description{
Print method for greta python deps
}
