% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R, R/install_tensorflow.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{py_module_available}
\alias{py_version}
\alias{install_tensorflow}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{reticulate}{\code{\link[reticulate]{py_module_available}}, \code{\link[reticulate]{py_version}}}

  \item{tensorflow}{\code{\link[tensorflow]{install_tensorflow}}}
}}

