% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_tensorflow.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{install_tensorflow}
\alias{window}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{stats}{\code{\link[stats]{window}}}

  \item{tensorflow}{\code{\link[tensorflow]{install_tensorflow}}}
}}

