% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimiser_class.R
\name{run_optimiser}
\alias{run_optimiser}
\title{Dispatching the optimisation method based on the class.}
\usage{
run_optimiser(self)
}
\arguments{
\item{self}{Object of class "tf_optimiser", "tfp_optimiser", or
"tf_compat_optimiser".}
}
\description{
Dispatching the optimisation method based on the class.
}
