% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greta_stash.R
\name{greta_notes_install_miniconda_output}
\alias{greta_notes_install_miniconda_output}
\alias{stash-notes}
\alias{greta_notes_install_miniconda_error}
\alias{greta_notes_conda_create_output}
\alias{greta_notes_conda_create_error}
\alias{greta_notes_conda_install_output}
\alias{greta_notes_conda_install_error}
\alias{greta_notes_tf_num_error}
\title{Retrieve python installation or error details.}
\usage{
greta_notes_install_miniconda_output()

greta_notes_install_miniconda_error()

greta_notes_conda_create_output()

greta_notes_conda_create_error()

greta_notes_conda_install_output()

greta_notes_conda_install_error()

greta_notes_tf_num_error()
}
\description{
These functions retrieve installation or error information
output by python when running \code{install_miniconda()}, \code{conda_create()},
\code{conda_install()}, or when encountering a TensorFlow numerical problem.
}
\examples{
\dontrun{
greta_notes_install_miniconda()
greta_notes_conda_create()
greta_notes_conda_install()
greta_notes_tf_num_error()
greta_notes_tf_error()
}
}
