% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greta_stash.R
\name{greta_notes_tf_num_error}
\alias{greta_notes_tf_num_error}
\title{Retrieve python messages.}
\usage{
greta_notes_tf_num_error()
}
\description{
These functions retrieve specific python error messages that might
come up during greta use.
}
\examples{
\dontrun{
greta_notes_tf_error()
}
}
