% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-logfiles.R
\name{write_greta_install_log}
\alias{write_greta_install_log}
\title{Write greta dependency installation log file}
\usage{
write_greta_install_log(path = greta_logfile)
}
\arguments{
\item{path}{a path with an HTML (.html) extension.}
}
\value{
nothing - writes to file
}
\description{
This can only be run after installation has happened with
\code{\link[=install_greta_deps]{install_greta_deps()}}, and before restarting R.
}
